requirements.txt
pyyaml>=6.0.1
jsonschema>=4.22.0
rich>=13.7.1
gitpython>=3.1.43
typing-extensions>=4.12.2
pytest>=8.2.0

PyGithub>=2.4.0
python-dotenv>=1.0.1
PyGithub
